consolidate test helper and utility
